Bug Fixes (#30)
* Bug Fixes

- Fixed nextPage() and previousPage() in Collections
- SimCard: Fixed register() function
- Fixed AddressTest
- Fixed CallControlApplicationTest
- Fixed CollectionTest
- Fixed PartoutTest

* Update All.php

Added 'url' to Collection::all() for nextPage() and previousPage()

* Updated Trait All

When calling ::all() we used to force the library to use record_type = 'list' to detect Telnyx\Collection. New behavior detects the presence of the 'meta' instead because 'meta' is only present for lists.

* Update Collection.php

Now passing params from last call back into the return object. Important for nextPage() and previousPage().